BE: Add logout with query parameters to generic OAuth provider
The logout URL is taken from auto-detected endpoint from the provider's
well known configuration address, but can be overridden by custom
parameter `logout-url`. Query parameters are optional and are added only
when the corresponding key is present. Default redirect URL is a base URL
and can be overridden in custom parameters.

Also the final slash from the redirect URL can be removed when
`redirect-url-strip-slash` is set to true/yes/on/1.

Example with all custom parameters:

```
custom-params:
  type: oauth
  logout-url: https://example.com/sso/logout
  client-id-key: clientId
  redirect-url-key: returnTo
  redirect-url: https://example.com/kafka-ui
  redirect-url-strip-slash: true
```

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>